Separate main view from subviews
